requirements.txt
# Web Framework & Server
fastapi==0.111.0
uvicorn[standard]==0.30.1

# AI Orchestration & Observability
langchain
langchain-core
langgraph
langsmith

# LLM Integrations
langchain-groq
langchain-community

# Configuration & Env Vars
pydantic-settings==2.2.1
python-dotenv==1.0.1


# Vector Database & Embeddings
langchain-chroma>=0.1.2
chromadb>=0.5.0

# Security & Validation
sqlglot==25.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Persistence
langgraph-checkpoint-sqlite>=1.0.0

# Observability
structlog==24.4.0

langchain-huggingface
sentence-transformers

PyPDF2
python-multipart